Now ntp correction is drifting by 0.5 seconds every 5 seconds, but I'm worried my code causes that number to go up or down- maybe hprintln inside interrupt_free wasn't helping.
Instead of any hprintln at all send a timestamp packet to the computer that contains the ntp information, and time how long between packets- the stm32 thinks there is a 5 second delay but maybe it's 4.5 as measured from the outside.  Also how to convert current timestamp to the ntp corrected time?  I think there is a function within sntpc.

```
[5001] sntp offset: 1754524446.97s
[10001] sntp offset: 1754524447.57s
[15001] sntp offset: 1754524448.18s
[20001] sntp offset: 1754524448.79s
[25001] sntp offset: 1754524449.39s
[30001] sntp offset: 1754524449.99s
[35001] sntp offset: 1754524450.60s
```